# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "File Management System API",
    "description": "Backend for a complete file management system: JWT auth with email-verification OTPs, file uploads with content extraction (PDF/DOCX/XLSX/CSV/text/images) and Cloudinary storage, per-user dashboards, admin statistics, and admin user management.\n\nAll authenticated endpoints expect `Authorization: Bearer <accessToken>`. The token payload carries `{ userId, role }` only.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:8080", "description": "Local development" },
    { "url": "https://your-project.vercel.app", "description": "Vercel deployment (replace with your domain)" }
  ],
...
  "paths": {
...
    "/auth/verify-email": {
      "post": {
        "tags": ["Auth"],
        "summary": "Verify email with OTP",
        "description": "Consumes the newest active EMAIL_VERIFY code for the account and flips `isVerified` to true. Codes expire after 10 minutes; resending invalidates previous codes.",
        "operationId": "verifyEmail",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/VerifyEmailInput" },
              "example": { "email": "jane@example.com", "code": "482913" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email verified",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthUserResponse" },
                "example": {
                  "message": "Email verified successfully",
                  "user": {
                    "id": "9b8c1be2-07ba-4075-8c07-82e49ac505ea",
                    "name": "Jane Tester",
                    "email": "jane@example.com",
                    "role": "USER",
                    "isVerified": true,
                    "createdAt": "2026-08-24T22:31:18.633Z",
                    "updatedAt": "2026-08-24T22:31:27.070Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid, expired, or already-consumed code",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorEnvelope" },
                "example": { "error": { "code": "VALIDATION_ERROR", "message": "Invalid or expired verification code" } }
              }
            }
          },
          "409": {
            "description": "Email already verified",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorEnvelope" },
                "example": { "error": { "code": "CONFLICT", "message": "Email is already verified" } }
              }
            }
          }
        }
      }
    },
...
    "/auth/forgot-password": {
      "post": {
        "tags": ["Auth"],
        "summary": "Request a password reset code",
        "description": "Issues a PASSWORD_RESET OTP (10-minute validity) if the account exists. Response is generic always (anti-enumeration).",
        "operationId": "forgotPassword",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/EmailInput" },
              "example": { "email": "jane@example.com" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Generic success regardless of account existence",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageOnly" },
                "example": { "message": "If that email exists, a password reset code has been sent." }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" }
        }
      }
    },
    "/auth/reset-password": {
      "post": {
        "tags": ["Auth"],
        "summary": "Reset password with OTP",
        "description": "Verifies and consumes the newest PASSWORD_RESET code, then replaces the password. For unknown emails the error is identical to a bad code (anti-enumeration).",
        "operationId": "resetPassword",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ResetPasswordInput" },
              "example": { "email": "jane@example.com", "code": "731904", "newPassword": "N3wPass!456" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthUserResponse" },
                "example": {
                  "message": "Password reset successfully",
                  "user": {
                    "id": "9b8c1be2-07ba-4075-8c07-82e49ac505ea",
                    "name": "Jane Tester",
                    "email": "jane@example.com",
                    "role": "USER",
                    "isVerified": true,
                    "createdAt": "2026-08-24T22:31:18.633Z",
                    "updatedAt": "2026-08-24T22:31:48.633Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid/expired code or unknown email (indistinguishable)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorEnvelope" },
                "example": { "error": { "code": "VALIDATION_ERROR", "message": "Invalid or expired verification code" } }
              }
            }
          }
        }
      }
    },
    "/auth/change-password": {
      "post": {
        "tags": ["Auth"],
        "summary": "Change own password",
        "description": "Authenticated. Verifies the current password before applying the new one.",
        "operationId": "changePassword",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChangePasswordInput" },
              "example": { "currentPassword": "N3wPass!456", "newPassword": "Ch4nged!789" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthUserResponse" },
                "example": {
                  "message": "Password changed successfully",
                  "user": {
                    "id": "9b8c1be2-07ba-4075-8c07-82e49ac505ea",
                    "name": "Jane Tester",
                    "email": "jane@example.com",
                    "role": "USER",
                    "isVerified": true,
                    "createdAt": "2026-08-24T22:31:18.633Z",
                    "updatedAt": "2026-08-24T22:32:12.340Z"
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": {
            "description": "Missing/invalid token, or current password incorrect",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorEnvelope" },
                "examples": {
                  "wrongCurrent": { "value": { "error": { "code": "UNAUTHORIZED", "message": "Current password is incorrect" } } },
                  "noToken": { "value": { "error": { "code": "UNAUTHORIZED", "message": "Authentication required" } } }
                }
              }
            }
          }
        }
      }
    },
    "/auth/profile": {
      "get": {
        "tags": ["Auth"],
        "summary": "Get own profile",
        "description": "Authenticated. Returns the caller's public user object (password never included anywhere in this API).",
        "operationId": "getProfile",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Current user",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResponse" },
                "example": {
                  "user": {
                    "id": "9b8c1be2-07ba-4075-8c07-82e49ac505ea",
                    "name": "Jane Tester",
                    "email": "jane@example.com",
                    "role": "USER",
                    "isVerified": true,
                    "createdAt": "2026-08-24T22:31:18.633Z",
                    "updatedAt": "2026-08-24T22:31:48.633Z"
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": {
            "description": "Token references a deleted user",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorEnvelope" },
                "example": { "error": { "code": "NOT_FOUND", "message": "User not found" } }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token returned by POST /auth/login. Payload: { userId, role }."
      }
    },
    "schemas": {
...
      "Role": { "type": "string", "enum": ["USER", "ADMIN"] },
...
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string", "minLength": 1, "maxLength": 100 },
          "email": { "type": "string", "format": "email" },
          "role": { "$ref": "#/components/schemas/Role" },
          "isVerified": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "required": ["id", "name", "email", "role", "isVerified", "createdAt", "updatedAt"]
      },
...
      "VerifyEmailInput": {
        "type": "object",
        "required": ["email", "code"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "code": { "type": "string", "pattern": "^\\d{6}$", "description": "6-digit numeric code, 10-minute validity." }
        }
      },
      "EmailInput": {
        "type": "object",
        "required": ["email"],
        "properties": { "email": { "type": "string", "format": "email" } }
      },
      "ResetPasswordInput": {
        "type": "object",
        "required": ["email", "code", "newPassword"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "code": { "type": "string", "pattern": "^\\d{6}$" },
          "newPassword": { "type": "string", "minLength": 8, "maxLength": 72 }
        }
      },
      "ChangePasswordInput": {
        "type": "object",
        "required": ["currentPassword", "newPassword"],
        "properties": {
          "currentPassword": { "type": "string", "minLength": 1 },
          "newPassword": { "type": "string", "minLength": 8, "maxLength": 72 }
        }
      },
...
      "MessageOnly": {
        "type": "object",
        "properties": { "message": { "type": "string" } },
        "required": ["message"]
      },
      "UserResponse": {
        "type": "object",
        "properties": { "user": { "$ref": "#/components/schemas/User" } },
        "required": ["user"]
      },
      "AuthUserResponse": {
        "type": "object",
        "properties": {
          "message": { "type": "string" },
          "user": { "$ref": "#/components/schemas/User" }
        },
        "required": ["message", "user"]
      },
...
      "ErrorDetail": {
        "type": "object",
        "properties": {
          "path": { "type": "string" },
          "message": { "type": "string" }
        },
        "required": ["path", "message"]
      },
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": ["VALIDATION_ERROR", "UNAUTHORIZED", "FORBIDDEN", "NOT_FOUND", "CONFLICT", "FILE_TOO_LARGE", "MULTIPART_ERROR", "UPLOAD_FAILED", "EMAIL_SEND_FAILED", "INTERNAL_SERVER_ERROR"]
              },
              "message": { "type": "string" },
              "details": {
                "type": "array",
                "nullable": true,
                "items": { "$ref": "#/components/schemas/ErrorDetail" },
                "description": "Present only for zod validation failures."
              }
            },
            "required": ["code", "message"]
          }
        },
        "required": ["error"]
      }
    },
    "responses": {
      "ValidationError": {
        "description": "Request body failed validation",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" },
            "example": {
              "error": {
                "code": "VALIDATION_ERROR",
                "message": "Invalid request payload",
                "details": [
                  { "path": "name", "message": "Name is required" },
                  { "path": "email", "message": "Must be a valid email address" },
                  { "path": "password", "message": "Password must be at least 8 characters" }
                ]
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing, invalid, or expired bearer token",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" },
            "example": { "error": { "code": "UNAUTHORIZED", "message": "Authentication required" } }
          }
        }
      },
...
    }
  }
}